コメントなどの細かい部分を追加

+++ StageScene.h
@@ -9,8 +9,8 @@
 #include "Global.h"
 #include "GoalStaging.h"
 #include "HP.h"
-#include "ItemModel.h"
 #include "ItemImage.h"
+#include "ItemModel.h"
 #include "Pause.h"
 #include "Sign.h"
 #include "Stage.h"
@@ -27,9 +27,9 @@ class StageScene : public GameObject
 	int Pselect_; //Pause選択用変数
 
 	//定数
-	const int CAMERA_SET_POS = -10; //カメラ位置
-	const int CAMERA_FALL_MAX = -3; //落下時のカメラ限界
-	const int FALL_MAX = -8;        //Playerの落下限界
+	const float CAMERA_SET_POS = -10.f;  //カメラ位置
+	const float CAMERA_FALL_MAX = -3.f;  //落下時のカメラ限界
+	const int FALL_MAX = -8;             //Playerの落下限界
 	const XMFLOAT3 ITEMIMAGE_SCALE_STAGE = XMFLOAT3(0.5f,0.5f,0.5f);     //アイテム画像のサイズ
 	const XMFLOAT3 ITEMIMAGE_POSITION_STAGE = XMFLOAT3(0.9f, 0.8f, 0.f); //アイテム画像の位置
